style: change style

+++ src/components/TodoItem.vue
@@ -7,7 +7,7 @@
       </div>
     </div>
     <div class="todoItemButton">
-      <button @click="editMode = true" class="appButton isWarning">
+      <button @click="editMode = true" class="appButton isPrimary">
         編集
       </button>
       <button @click="deleteTodo" class="appButton isDanger">削除</button>
@@ -15,24 +15,30 @@
   </div>
   <div v-else class="todoItem">
     <form class="appForm">
-      <div class="formControl">
-        <label class="label">タイトル</label>
-        <input v-model="todo.title" class="formInput" type="text" />
-      </div>
-      <div class="formControl formControlLast">
-        <label class="label">メモ</label>
-        <textarea
-          v-model="todo.description"
-          cols="30"
-          rows="2"
-          class="formInput"
-        />
+      <div class="formControlWrapper">
+        <div class="formControl">
+          <input
+            placeholder="タイトル"
+            v-model="todo.title"
+            class="formInput textField"
+            type="text"
+          />
+        </div>
+        <div class="formControl formControlLast">
+          <textarea
+            placeholder="メモ"
+            v-model="todo.description"
+            cols="30"
+            rows="2"
+            class="formInput textField"
+          />
+        </div>
       </div>
       <div class="todoItemButton">
-        <button @click.prevent="updateTodo" class="appButton isWarning">
-          更新する
+        <button @click.prevent="updateTodo" class="appButton isPrimary">
+          完了
         </button>
-        <button @click.prevent="editMode = false" class="appButton isCancel">
+        <button @click.prevent="editMode = false" class="appButton isSecondary">
           キャンセル
         </button>
       </div>
@@ -71,7 +77,10 @@ export default {
       this.editMode = false;
     },
     deleteTodo() {
-      store.dispatch("deleteTodo", this.todo._id);
+      const result = window.confirm("削除します。よろしいですか？");
+      if (result) {
+        store.dispatch("deleteTodo", this.todo._id);
+      }
     },
   },
 };
@@ -80,20 +89,21 @@ export default {
 <style lang="scss" scoped>
 .todo {
   &Item {
-    background-color: gray;
+    background-color: #fff;
+    border: 2px solid #fafafa;
     margin: 10px;
-    padding: 10px;
-    color: white;
-    border-radius: 5px;
+    padding: 16px;
+    border-radius: 16px;
     font-size: 23px;
+    box-shadow: 0 2px 4px rgba(67, 133, 187, 0.07);
     &Content {
       margin-bottom: 5px;
       &Title {
-        font-size: 24px;
+        font-size: 20px;
         font-weight: bold;
       }
       &Description {
-        font-size: 16px;
+        font-size: 14px;
       }
     }
     &Button {
@@ -102,7 +112,7 @@ export default {
       justify-content: flex-start;
       .appButton {
         font-size: 15px;
-        margin-right: 5px;
+        margin-right: 8px;
       }
     }
   }
